+++ redalert/session.cpp
@@ -1056,7 +1056,7 @@ unsigned long SessionClass::Compute_Unique_ID(void)
     return 1; // PG
 #if (0)       // PG
     time_t tm;
-    unsigned long id;
+    unsigned int id;
     struct diskfree_t dtable;
     char* path;
     int i;
